Time: 144 ms (92.18%), Space: 71.1 MB (10.29%) - LeetHub

+++ 3332-minimum-operations-to-exceed-threshold-value-ii/3332-minimum-operations-to-exceed-threshold-value-ii.java
@@ -1,12 +1,12 @@
 class Solution {
     public int minOperations(int[] nums, int k) {
         int ans = 0;
-        Queue<Double> q = new PriorityQueue<>();
+        Queue<Long> q = new PriorityQueue<>();
 
         for(int num : nums) {
-            q.offer((double)num);
+            q.offer((long)num);
         }
-        double a,b;
+        long a,b;
         while(q.size() > 0) {
             a = q.poll();
             if(a >= k) break;
